0202 sujin pm05:53

+++ own/src/main/java/co/prjt/own/chall/service/CResultVO.java
@@ -12,7 +12,7 @@ public class CResultVO {
 	int memVld; // 회원이 완료한 인증횟수
 	int challVld; //도전 인증횟수라 숫자
 	
-	int successRate; //횟수를 통해서 계산한 도전성공률
+	double successRate; //횟수를 통해서 계산한 도전성공률
 	int refundPrice; //도전환급금
 	int rewardPrice; //100% 달성시 받는 상금
 }
